+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 ## v4.0.0 (unreleased)
+
+### Features
+- added `isOnNewgrounds()` method for check if the game is current running on Newgrounds.com
+
+### Misc
 - fully rewrite of the library
 - now using fetch api (async/await support in all methods)
 - TSDoc for all methods, responses and calls